Dialog implemented

+++ app/src/main/java/com/example/clauditter/ui/Favorites/FavoritesFragment.kt
@@ -1,5 +1,6 @@
 package com.example.clauditter.ui.Favorites
 
+import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -26,11 +27,11 @@ import kotlinx.android.synthetic.main.fragment__trailers.*
 import kotlinx.android.synthetic.main.fragment_favorites.*
 
 
-
 class FavoritesFragment : Fragment(),
-    OnRecyclerClickListener{
+    OnRecyclerClickListener {
 
     private val logInModel: ViewModel_LogIn by activityViewModels()
+
     /**RECYCLER ADAPTER */
     private var favoritesAdapter: FavoritesAdapter = FavoritesAdapter(ArrayList())
 
@@ -40,17 +41,17 @@ class FavoritesFragment : Fragment(),
         savedInstanceState: Bundle?
     ): View? {
         val root = inflater.inflate(R.layout.fragment_favorites, container, false)
-       //show/hide elements if the user is logged in
+        //show/hide elements if the user is logged in
         logInModel.flag.observe(viewLifecycleOwner, Observer {
             if (it) {
                 lbl_warningFavorites.visibility = View.GONE
                 recycler_favorites.visibility = View.VISIBLE
                 lbl_favoritesMessage.visibility = View.VISIBLE
-               // if (logInModel.favoriteListIsLoaded()) {//downloading new data,
+                // if (logInModel.favoriteListIsLoaded()) {//downloading new data,
                 /*This is commented because if not, the list wouldn`t be always updated*/
-                    Toast.makeText(activity, "downloading new data", Toast.LENGTH_LONG).show()
-                    getFavorites()
-            //    }
+                Toast.makeText(activity, "downloading new data", Toast.LENGTH_LONG).show()
+                getFavorites()
+                //    }
             } else {
                 lbl_warningFavorites.visibility = View.VISIBLE
                 recycler_favorites.visibility = View.GONE
@@ -60,7 +61,7 @@ class FavoritesFragment : Fragment(),
 
         logInModel.user.observe(viewLifecycleOwner, Observer {
             lbl_favoritesMessage.text = "$it´ s Favorites List"
-           // getFavorites() //new user, new favorites, this is not necessary anymore because the if above is comented
+            // getFavorites() //new user, new favorites, this is not necessary anymore because the if above is comented
         })
 
 
@@ -74,7 +75,13 @@ class FavoritesFragment : Fragment(),
         super.onViewCreated(view, savedInstanceState)
         recycler_favorites.layoutManager = LinearLayoutManager(activity)
         recycler_favorites.adapter = favoritesAdapter
-        recycler_favorites.addOnItemTouchListener(RecyclerItemsListeners(view.context,recycler_favorites,this))
+        recycler_favorites.addOnItemTouchListener(
+            RecyclerItemsListeners(
+                view.context,
+                recycler_favorites,
+                this
+            )
+        )
     }
 
     private fun getFavorites() {
@@ -100,7 +107,7 @@ class FavoritesFragment : Fragment(),
                                 document["movieId"].toString().toInt(),
                                 document["movieTitle"].toString(),
                                 document["photo"].toString(),
-                                        document.id
+                                document.id
                             )
                         )
                     }
@@ -112,39 +119,37 @@ class FavoritesFragment : Fragment(),
 
     /**Listeners for the recyclerView Items*/
     override fun onItemClick(view: View, position: Int) {
-        val favMovie=favoritesAdapter.getFavorite(position)
-        val movie=logInModel.getMovie(favMovie?.movieid)
-        if(movie!=null){
-            Toast.makeText(view.context ," ${movie.title}", Toast.LENGTH_SHORT).show()
-            val intent= Intent(view.context, Activity_MovieDetails::class.java)
-            intent.putExtra(MOVIE_TRANSFER,movie)
-            intent.putExtra(USERNAME,logInModel.user.value)
-            intent.putExtra(IS_LOGED,logInModel.flag.value)
+        val favMovie = favoritesAdapter.getFavorite(position)
+        val movie = logInModel.getMovie(favMovie?.movieid)
+        if (movie != null) {
+            Toast.makeText(view.context, " ${movie.title}", Toast.LENGTH_SHORT).show()
+            val intent = Intent(view.context, Activity_MovieDetails::class.java)
+            intent.putExtra(MOVIE_TRANSFER, movie)
+            intent.putExtra(USERNAME, logInModel.user.value)
+            intent.putExtra(IS_LOGED, logInModel.flag.value)
             view.context.startActivity(intent)
         }
     }
 
     override fun onItemLongClick(view: View, position: Int) {
-            //todo implement confirmation dialog
-        val favMovie=favoritesAdapter.getFavorite(position)
-               if(favMovie!=null){
-                   val db = FirebaseFirestore.getInstance()
-                   db.collection("favorites").document(favMovie.id)
-                       .delete()
-                       .addOnSuccessListener { Toast.makeText(
-                           view.context,
-                           "${favMovie.movieTitle} deleted from  favorites",
-                           Toast.LENGTH_SHORT
-                       ).show()
-
-                           }
-                       .addOnFailureListener { Toast.makeText(
-                           view.context,
-                           "There was an error :(",
-                           Toast.LENGTH_SHORT
-                       ).show() }
-
-
-               }
+        val favMovie = favoritesAdapter.getFavorite(position)
+        if (favMovie != null) {
+            val builder = AlertDialog.Builder(activity)
+            builder.setTitle("Delete From Favorite")
+            builder.setMessage("Are you sure you want to delete ${favMovie.movieTitle}")
+            builder.setPositiveButton(android.R.string.yes) { dialog, which ->
+                val db = FirebaseFirestore.getInstance()
+                db.collection("favorites").document(favMovie.id)
+                    .delete()
+                    .addOnSuccessListener {
+                        Toast.makeText(view.context, "${favMovie.movieTitle} deleted from  favorites", Toast.LENGTH_SHORT).show()
+                        dialog.cancel() }
+                    .addOnFailureListener { Toast.makeText(view.context, "There was an error :(", Toast.LENGTH_SHORT).show() }
+            }
+            builder.setNegativeButton(android.R.string.no) { dialog, which ->
+                dialog.cancel()
+            }
+            builder.show()
+        }
     }
 }//End Of Class
